+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [1.11.12] - unreleased
 ### Added
 ### Changed
+- `ghorg reclone --list` now returns a sorted list
 ### Deprecated
 ### Removed
 ### Fixed